refactor: move myflags into own file
